Added Rounding + Strings are now defined in srings.xml
Double is now displayed as Int

+++ app/src/main/java/com/example/bmweather/WeatherResponse.kt
@@ -59,7 +59,7 @@ class Wind {
 
 class Main {
     @SerializedName("temp")
-    var temp: Float = 0.toFloat()
+    var temp: Double = 0.0
     @SerializedName("humidity")
     var humidity: Float = 0.toFloat()
     @SerializedName("feels_like")
